Found colours
I tested different colours for the shapes and background and made sure they worked well together.
It's important they are distinct enough to stack because I plan on using a strokeless style, so the shapes need to be easily distinguished.
I also have them all in different push, pop statements, which could be helpful later when I start actually drawing.