Added in some dialog for the frog to say later on
I want the frog to talk to the player.
Because the old people I know like to talk to others. Even if they're a bit shy at first.

I will be using random to randomly display dialogs
So someone might get unlucky and get the same one for the entire duration of the game. But what'sthe probability of that... right?