Tutorial implementation part1
Started writing the tutorial
Forgot to commit to show progress but I made the first part of the tutorial, explaining frog movement and fly catching.

I made a function to draw textboxes so that saves me lots of time and trouble and makes the code much more condensed.
I'm using a switch to control the tutorial. Not sure if that's the most efficient method since I am still copy pasting code but it helps me visualize the different steps users go through in my tutorial.

Next will be introducing the math fly.