Input field!
So turns out it was reading the width of my entire screen and not just the browser window.

it's fixed now!

Input is at the correct spot unless you resize the window during gameplay...

Maybe if i have time i can fix that by checking if the screen size changed between frames.

For now i want to style the input field, so I'll be using css!