There's a little wizard!
Made a little wizard, her name is Sumi.

I'll need to organize all the values into an object and keep track of the mouth so I can sort of animate her.

I also need to put the UIOffset on her so she shakes with the rest of the screen.

Finally, her speech bubble, she needs to guide the user through the game!