Sounds work!
There are now some little sounds to give audio feedback and a cute little music loop!

Unfortunately, the sounds don't play over each other so they wait for the other sound to stop playing before playing again (which is not so good for players who like to spam)

I'm working on the title page now!